+++ src/Products/CMFPlomino/defaultmacros/macros/macro_condition_field_got_substring.formula.py
@@ -67,3 +67,25 @@ return code
 
 
 ## END formula }
+## START document_title {
+doc = plominoContext
+field_name_ = doc.getItem('field_name', '')
+field_compare_ = doc.getItem('field_compare', '')
+field_value_ = doc.getItem('field_value', '')
+value_type_ = doc.getItem('value_type', '')
+code = 'Field contains text'
+
+if not (field_name_ and field_compare_ and field_value_ and value_type_):
+    return code
+
+value_type = 'pattern'
+if value_type_ == 'text':
+    value_type = ''
+code = '"{field_name}" {field_compare} {field_value} {value_type}'.format(
+    field_name=field_name_,
+    field_compare=field_compare_,
+    field_value=field_value_,
+    value_type=value_type
+)
+return code
+## END document_title }